refactor: rename widgets app tester files
