+++ packages/flutter/test/widgets/gesture_detector_test.dart
@@ -7,7 +7,7 @@ import 'package:flutter/rendering.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import 'test_widgets_app.dart';
+import 'widgets_app_tester.dart';
 
 void main() {
   const forcePressOffset = Offset(400.0, 50.0);
