+++ packages/flutter/test/widgets/opacity_test.dart
@@ -14,7 +14,7 @@ import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import 'semantics_tester.dart';
-import 'test_widgets_app.dart';
+import 'widgets_app_tester.dart';
 
 void main() {
   testWidgets('Opacity', (WidgetTester tester) async {
