+++ packages/flutter/test/widgets/widgets_app_tester_test.dart
@@ -5,7 +5,7 @@
 import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import 'test_widgets_app.dart';
+import 'widgets_app_tester.dart';
 
 void main() {
   group('TestWidgetsApp', () {
